split settings modals
