+++ src/app/app.less
@@ -664,3 +664,161 @@ a.a-block {
         }
     }
 }
+
+.settings-field {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+
+    &.settings-field.sub-field {
+        .settings-label {
+            font-weight: normal;
+
+            text-align: right;
+            padding-right: 20px;
+        }
+    }
+
+    &.settings-error {
+        color: @term-red;
+        margin-top: 20px;
+        padding: 10px;
+        border-radius: 5px;
+        background-color: #200707;
+        border: 1px solid @term-red;
+        font-weight: bold;
+
+        .error-dismiss {
+            padding: 2px;
+            cursor: pointer;
+        }
+    }
+
+    .settings-label {
+        font-weight: bold;
+        width: 12em;
+        display: flex;
+        flex-direction: row;
+        align-items: center;
+        .info-message {
+            margin-left: 5px;
+        }
+    }
+
+    .settings-input {
+        display: flex;
+        flex-direction: row;
+        align-items: center;
+        color: @term-white;
+
+        &.settings-clickable {
+            cursor: pointer;
+        }
+
+        &.inline-edit.edit-active {
+            input.input {
+                padding: 0;
+                height: 20px;
+            }
+
+            .button {
+                height: 20px;
+            }
+        }
+
+        input {
+            padding: 4px;
+            border-radius: 3px;
+        }
+
+        .control {
+            .icon {
+                width: 1.5em;
+                height: 1.5em;
+                margin: 0;
+            }
+        }
+
+        .tab-color-icon.color-default path {
+            fill: @tab-green;
+        }
+        .tab-color-icon.color-green path {
+            fill: @tab-green;
+        }
+        .tab-color-icon.color-orange path {
+            fill: @tab-orange;
+        }
+        .tab-color-icon.color-red path {
+            fill: @tab-red;
+        }
+        .tab-color-icon.color-yellow path {
+            fill: @tab-yellow;
+        }
+        .tab-color-icon.color-blue path {
+            fill: @tab-blue;
+        }
+        .tab-color-icon.color-mint path {
+            fill: @tab-mint;
+        }
+        .tab-color-icon.color-cyan path {
+            fill: @tab-cyan;
+        }
+        .tab-color-icon.color-white path {
+            fill: @tab-white;
+        }
+        .tab-color-icon.color-violet path {
+            fill: @tab-violet;
+        }
+        .tab-color-icon.color-pink path {
+            fill: @tab-pink;
+        }
+
+        .tab-colors,
+        .tab-icons {
+            display: flex;
+            flex-direction: row;
+            align-items: center;
+
+            .tab-color-sep,
+            .tab-icon-sep {
+                padding-left: 10px;
+                padding-right: 10px;
+                font-weight: bold;
+            }
+
+            .tab-color-icon,
+            .tab-icon-icon {
+                width: 1.1em;
+                vertical-align: middle;
+            }
+
+            .tab-color-name,
+            .tab-icon-name {
+                margin-left: 1em;
+            }
+
+            .tab-color-select,
+            .tab-icon-select {
+                cursor: pointer;
+                margin: 5px;
+                &:hover {
+                    outline: 2px solid white;
+                }
+            }
+        }
+
+        .action-text {
+            margin-left: 20px;
+
+            color: @term-red;
+        }
+
+        .settings-share-link {
+            width: 160px;
+        }
+    }
+
+    &:not(:first-child) {
+        margin-top: 10px;
+    }
+}
