+++ src/app/common/modals/index.tsx
@@ -6,3 +6,6 @@ export { CreateRemoteConnModal } from "./createremoteconn";
 export { ViewRemoteConnDetailModal } from "./viewremoteconndetail";
 export { EditRemoteConnModal } from "./editremoteconn";
 export { TabSwitcherModal } from "./tabswitcher";
+export { SessionSettingsModal } from "./sessionsettings";
+export { ScreenSettingsModal } from "./screensettings";
+export { LineSettingsModal } from "./linesettings";
