+++ src/app/common/modals/linesettings.less
@@ -1,0 +1,23 @@
+@import "../../../app/common/themes/themes.less";
+
+.line-settings-modal {
+    width: 640px;
+
+    .wave-modal-content {
+        gap: 24px;
+
+        .wave-modal-body {
+            display: flex;
+            padding: 0px 20px;
+            flex-direction: column;
+            align-items: flex-start;
+            gap: 4px;
+            align-self: stretch;
+            width: 100%;
+
+            .settings-input .hotkey {
+                color: @text-secondary;
+            }
+        }
+    }
+}
