+++ src/app/common/modals/sessionsettings.less
@@ -1,0 +1,19 @@
+@import "../../../app/common/themes/themes.less";
+
+.session-settings-modal {
+    width: 640px;
+
+    .wave-modal-content {
+        gap: 24px;
+
+        .wave-modal-body {
+            display: flex;
+            padding: 0px 20px;
+            flex-direction: column;
+            align-items: flex-start;
+            gap: 4px;
+            align-self: stretch;
+            width: 100%;
+        }
+    }
+}
